+++ src/compiler/Disassembler.h
@@ -36,7 +36,7 @@ class Disassembler {
 public:
     static void Dump(pVMClass cl);
     static void DumpMethod(pVMMethod method, const char* indent);
-    static void DumpBytecode(pVMFrame frame, pVMMethod method, int bc_idx);
+    static void DumpBytecode(pVMFrame frame, pVMMethod method, long bc_idx);
 private:
     static void dispatch(pVMObject o);
 };
