+++ src/vmobjects/VMInvokable.h
@@ -36,7 +36,7 @@ class VMFrame;
 
 class VMInvokable: public VMObject {
 public:
-    VMInvokable(int nof = 0) : VMObject(nof + 2) {};
+    VMInvokable(long nof = 0) : VMObject(nof + 2) {};
     // virtual operator "()" to invoke the invokable
     virtual void      operator()(pVMFrame) = 0;
